removed legacy svgs and moved /home page to /
