+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "tailwindcss": "^3.4.6"
+  }
+}
