+++ src/interface/web/app/agents/page.tsx
@@ -72,7 +72,7 @@ async function openChat(slug: string, userData: UserProfile | null) {
 
     const response = await fetch(`/api/chat/sessions?agent_slug=${slug}`, { method: "POST" });
     if (response.status == 200) {
-        window.location.href = `/home`;
+        window.location.href = `/`;
     } else if (response.status == 403 || response.status == 401) {
         window.location.href = unauthenticatedRedirectUrl;
     } else {
