+++ src/interface/web/package.json
@@ -43,12 +43,12 @@
     "@types/react-dom": "^18",
     "autoprefixer": "^10.4.19",
     "class-variance-authority": "^0.7.0",
+    "clsx": "^2.1.1",
     "cmdk": "^1.0.0",
+    "cronstrue": "^2.50.0",
     "dompurify": "^3.1.6",
     "eslint": "^8",
     "eslint-config-next": "14.2.3",
-    "clsx": "^2.1.1",
-    "cronstrue": "^2.50.0",
     "katex": "^0.16.10",
     "lucide-react": "^0.397.0",
     "markdown-it": "^14.1.0",
@@ -62,9 +62,8 @@
     "shadcn-ui": "^0.8.0",
     "swr": "^2.2.5",
     "tailwind-merge": "^2.3.0",
-    "tailwindcss": "^3.4.4",
-    "typescript": "^5",
     "tailwindcss-animate": "^1.0.7",
+    "typescript": "^5",
     "vaul": "^0.9.1",
     "zod": "^3.23.8"
   },
@@ -80,6 +79,7 @@
     "lint-staged": "^15.2.7",
     "nodemon": "^3.1.3",
     "prettier": "3.3.3",
+    "tailwindcss": "^3.4.6",
     "typescript": "^5"
   },
   "prettier": {
